+++ src/BInterestRateModel.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: MIT
 pragma solidity 0.8.13;
 
 import {BStorage} from "./BStorage.sol";
@@ -8,50 +9,51 @@ import {PoolToken} from "./PoolToken.sol";
 /// @notice Interest rate model for Borrowable tokens.
 
 contract BInterestRateModel is PoolToken, BStorage {
-    // When utilization is 100% borrowRate is kinkBorrowRate * KINK_MULTIPLIER
-    // kinkBorrowRate relative adjustment per second belongs to [1-adjustSpeed, 1+adjustSpeed*(KINK_MULTIPLIER-1)]
-    uint public constant KINK_MULTIPLIER = 2;
-    uint public constant KINK_BORROW_RATE_MAX = 792.744800e9; //2500% per year
-    uint public constant KINK_BORROW_RATE_MIN = 0.31709792e9; //1% per year
+    /// @dev When utilization is 100% borrowRate is kinkBorrowRate * KINK_MULTIPLIER
+    /// @dev kinkBorrowRate relative adjustment per second belongs to [1-adjustSpeed, 1+adjustSpeed*(KINK_MULTIPLIER-1)]
+    uint256 public constant KINK_MULTIPLIER = 2;
+    uint256 public constant KINK_BORROW_RATE_MAX = 792.744800e9; //2500% per year
+    uint256 public constant KINK_BORROW_RATE_MIN = 0.31709792e9; //1% per year
 
     event AccrueInterest(
-        uint interestAccumulated,
-        uint borrowIndex,
-        uint totalBorrows
+        uint256 interestAccumulated,
+        uint256 borrowIndex,
+        uint256 totalBorrows
     );
-    event CalculateKinkBorrowRate(uint kinkBorrowRate);
-    event CalculateBorrowRate(uint borrowRate);
+    event CalculateKinkBorrowRate(uint256 kinkBorrowRate);
+    event CalculateBorrowRate(uint256 borrowRate);
 
     function _calculateBorrowRate() internal {
-        uint _kinkUtilizationRate = kinkUtilizationRate;
-        uint _adjustSpeed = adjustSpeed;
-        uint _borrowRate = borrowRate;
-        uint _kinkBorrowRate = kinkBorrowRate;
+        uint256 _kinkUtilizationRate = kinkUtilizationRate;
+        uint256 _adjustSpeed = adjustSpeed;
+        uint256 _borrowRate = borrowRate;
+        uint256 _kinkBorrowRate = kinkBorrowRate;
         uint32 _rateUpdateTimestamp = rateUpdateTimestamp;
 
-        // update kinkBorrowRate using previous borrowRate
-        uint32 timeElapsed = getBlockTimestamp() - _rateUpdateTimestamp; // underflow is desired
+        /// @dev update kinkBorrowRate using previous borrowRate
+        /// @dev underflow is desired
+        uint32 timeElapsed = getBlockTimestamp() - _rateUpdateTimestamp;
         if (timeElapsed > 0) {
             rateUpdateTimestamp = getBlockTimestamp();
-            uint adjustFactor;
+            uint256 adjustFactor;
 
             if (_borrowRate < _kinkBorrowRate) {
-                // never overflows, _kinkBorrowRate is never 0
-                uint tmp = ((((_kinkBorrowRate - _borrowRate) * 1e18) /
+                /// @dev never overflows, _kinkBorrowRate is never 0
+                uint256 tmp = ((((_kinkBorrowRate - _borrowRate) * 1e18) /
                     _kinkBorrowRate) *
                     _adjustSpeed *
                     timeElapsed) / 1e18;
                 adjustFactor = tmp > 1e18 ? 0 : 1e18 - tmp;
             } else {
-                // never overflows, _kinkBorrowRate is never 0
-                uint tmp = ((((_borrowRate - _kinkBorrowRate) * 1e18) /
+                /// @dev never overflows, _kinkBorrowRate is never 0
+                uint256 tmp = ((((_borrowRate - _kinkBorrowRate) * 1e18) /
                     _kinkBorrowRate) *
                     _adjustSpeed *
                     timeElapsed) / 1e18;
                 adjustFactor = tmp + 1e18;
             }
 
-            // never overflows
+            /// @dev never overflows
             _kinkBorrowRate = (_kinkBorrowRate * adjustFactor) / 1e18;
             if (_kinkBorrowRate > KINK_BORROW_RATE_MAX)
                 _kinkBorrowRate = KINK_BORROW_RATE_MAX;
@@ -62,27 +64,27 @@ contract BInterestRateModel is PoolToken, BStorage {
             emit CalculateKinkBorrowRate(_kinkBorrowRate);
         }
 
-        uint _utilizationRate;
+        uint256 _utilizationRate;
         {
-            // avoid stack to deep
-            uint _totalBorrows = totalBorrows; // gas savings
-            uint _actualBalance = totalBalance + _totalBorrows;
+            /// @dev avoid stack to deep
+            uint256 _totalBorrows = totalBorrows; // gas savings
+            uint256 _actualBalance = totalBalance + _totalBorrows;
             _utilizationRate = (_actualBalance == 0)
                 ? 0
                 : (_totalBorrows * 1e18) / _actualBalance;
         }
 
-        // update borrowRate using the new kinkBorrowRate
+        /// @dev update borrowRate using the new kinkBorrowRate
         if (_utilizationRate <= _kinkUtilizationRate) {
-            // never overflows, _kinkUtilizationRate is never 0
+            /// @dev never overflows, _kinkUtilizationRate is never 0
             _borrowRate =
                 (_kinkBorrowRate * _utilizationRate) /
                 _kinkUtilizationRate;
         } else {
-            // never overflows, _kinkUtilizationRate is always < 1e18
-            uint overUtilization = ((_utilizationRate - _kinkUtilizationRate) *
-                1e18) / (1e18 - _kinkUtilizationRate);
-            // never overflows
+            /// @dev never overflows, _kinkUtilizationRate is always < 1e18
+            uint256 overUtilization = ((_utilizationRate -
+                _kinkUtilizationRate) * 1e18) / (1e18 - _kinkUtilizationRate);
+            /// @dev never overflows
             _borrowRate =
                 (((KINK_MULTIPLIER - 1) * overUtilization + 1e18) *
                     _kinkBorrowRate) /
@@ -92,19 +94,21 @@ contract BInterestRateModel is PoolToken, BStorage {
         emit CalculateBorrowRate(_borrowRate);
     }
 
-    // applies accrued interest to total borrows and reserves
+    /// @notice applies accrued interest to total borrows and reserves
     function accrueInterest() public {
-        uint _borrowIndex = borrowIndex;
-        uint _totalBorrows = totalBorrows;
+        uint256 _borrowIndex = borrowIndex;
+        uint256 _totalBorrows = totalBorrows;
         uint32 _accrualTimestamp = accrualTimestamp;
 
         uint32 blockTimestamp = getBlockTimestamp();
+        /// @dev if same timestamp, terminate
         if (_accrualTimestamp == blockTimestamp) return;
-        uint32 timeElapsed = blockTimestamp - _accrualTimestamp; // underflow is desired
+        /// @dev underflow is desired
+        uint32 timeElapsed = blockTimestamp - _accrualTimestamp;
         accrualTimestamp = blockTimestamp;
 
-        uint interestFactor = uint(borrowRate) * timeElapsed;
-        uint interestAccumulated = (interestFactor * _totalBorrows) / 1e18;
+        uint256 interestFactor = uint256(borrowRate) * timeElapsed;
+        uint256 interestAccumulated = (interestFactor * _totalBorrows) / 1e18;
         _totalBorrows = _totalBorrows + interestAccumulated;
         _borrowIndex = _borrowIndex + ((interestFactor * _borrowIndex) / 1e18);
 
@@ -117,4 +121,3 @@ contract BInterestRateModel is PoolToken, BStorage {
         return uint32(block.timestamp % 2 ** 32);
     }
 }
-
