+++ src/interfaces/IBorrowable.sol
@@ -4,8 +4,8 @@ interface IBorrowable {
 
 	/*** Impermax ERC20 ***/
 	
-	event Transfer(address indexed from, address indexed to, uint value);
-	event Approval(address indexed owner, address indexed spender, uint value);
+	event Transfer(address indexed from, address indexed to, uint256 value);
+	event Approval(address indexed owner, address indexed spender, uint256 value);
 	
 	function name() external pure returns (string memory);
 	function symbol() external pure returns (string memory);
@@ -13,20 +13,20 @@ interface IBorrowable {
 	function totalSupply() external view returns (uint);
 	function balanceOf(address owner) external view returns (uint);
 	function allowance(address owner, address spender) external view returns (uint);
-	function approve(address spender, uint value) external returns (bool);
-	function transfer(address to, uint value) external returns (bool);
-	function transferFrom(address from, address to, uint value) external returns (bool);
+	function approve(address spender, uint256 value) external returns (bool);
+	function transfer(address to, uint256 value) external returns (bool);
+	function transferFrom(address from, address to, uint256 value) external returns (bool);
 	
 	function DOMAIN_SEPARATOR() external view returns (bytes32);
 	function PERMIT_TYPEHASH() external pure returns (bytes32);
 	function nonces(address owner) external view returns (uint);
-	function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+	function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s) external;
 	
 	/*** Pool Token ***/
 	
-	event Mint(address indexed sender, address indexed minter, uint mintAmount, uint mintTokens);
-	event Redeem(address indexed sender, address indexed redeemer, uint redeemAmount, uint redeemTokens);
-	event Sync(uint totalBalance);
+	event Mint(address indexed sender, address indexed minter, uint256 mintAmount, uint256 mintTokens);
+	event Redeem(address indexed sender, address indexed redeemer, uint256 redeemAmount, uint256 redeemTokens);
+	event Sync(uint256 totalBalance);
 	
 	function underlying() external view returns (address);
 	function factory() external view returns (address);
@@ -34,8 +34,8 @@ interface IBorrowable {
 	function MINIMUM_LIQUIDITY() external pure returns (uint);
 
 	function exchangeRate() external returns (uint);
-	function mint(address minter) external returns (uint mintTokens);
-	function redeem(address redeemer) external returns (uint redeemAmount);
+	function mint(address minter) external returns (uint256 mintTokens);
+	function redeem(address redeemer) external returns (uint256 redeemAmount);
 	function skim(address to) external;
 	function sync() external;
 	
@@ -43,9 +43,9 @@ interface IBorrowable {
 	
 	/*** Borrowable ***/
 
-	event BorrowApproval(address indexed owner, address indexed spender, uint value);
-	event Borrow(address indexed sender, address indexed borrower, address indexed receiver, uint borrowAmount, uint repayAmount, uint accountBorrowsPrior, uint accountBorrows, uint totalBorrows);
-	event Liquidate(address indexed sender, address indexed borrower, address indexed liquidator, uint seizeTokens, uint repayAmount, uint accountBorrowsPrior, uint accountBorrows, uint totalBorrows);
+	event BorrowApproval(address indexed owner, address indexed spender, uint256 value);
+	event Borrow(address indexed sender, address indexed borrower, address indexed receiver, uint256 borrowAmount, uint256 repayAmount, uint256 accountBorrowsPrior, uint256 accountBorrows, uint256 totalBorrows);
+	event Liquidate(address indexed sender, address indexed borrower, address indexed liquidator, uint256 seizeTokens, uint256 repayAmount, uint256 accountBorrowsPrior, uint256 accountBorrows, uint256 totalBorrows);
 	
 	function BORROW_FEE() external pure returns (uint);
 	function collateral() external view returns (address);
@@ -59,16 +59,16 @@ interface IBorrowable {
 	
 	function BORROW_PERMIT_TYPEHASH() external pure returns (bytes32);
 	function borrowApprove(address spender, uint256 value) external returns (bool);
-	function borrowPermit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
-	function borrow(address borrower, address receiver, uint borrowAmount, bytes calldata data) external;
-	function liquidate(address borrower, address liquidator) external returns (uint seizeTokens);
+	function borrowPermit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s) external;
+	function borrow(address borrower, address receiver, uint256 borrowAmount, bytes calldata data) external;
+	function liquidate(address borrower, address liquidator) external returns (uint256 seizeTokens);
 	function trackBorrow(address borrower) external;
 	
 	/*** Borrowable Interest Rate Model ***/
 
-	event AccrueInterest(uint interestAccumulated, uint borrowIndex, uint totalBorrows);
-	event CalculateKink(uint kinkRate);
-	event CalculateBorrowRate(uint borrowRate);
+	event AccrueInterest(uint256 interestAccumulated, uint256 borrowIndex, uint256 totalBorrows);
+	event CalculateKink(uint256 kinkRate);
+	event CalculateBorrowRate(uint256 borrowRate);
 	
 	function KINK_BORROW_RATE_MAX() external pure returns (uint);
 	function KINK_BORROW_RATE_MIN() external pure returns (uint);
@@ -84,9 +84,9 @@ interface IBorrowable {
 	
 	/*** Borrowable Setter ***/
 
-	event NewReserveFactor(uint newReserveFactor);
-	event NewKinkUtilizationRate(uint newKinkUtilizationRate);
-	event NewAdjustSpeed(uint newAdjustSpeed);
+	event NewReserveFactor(uint256 newReserveFactor);
+	event NewKinkUtilizationRate(uint256 newKinkUtilizationRate);
+	event NewAdjustSpeed(uint256 newAdjustSpeed);
 	event NewBorrowTracker(address newBorrowTracker);
 
 	function RESERVE_FACTOR_MAX() external pure returns (uint);
@@ -101,8 +101,8 @@ interface IBorrowable {
 		address _underlying, 
 		address _collateral
 	) external;
-	function _setReserveFactor(uint newReserveFactor) external;
-	function _setKinkUtilizationRate(uint newKinkUtilizationRate) external;
-	function _setAdjustSpeed(uint newAdjustSpeed) external;
+	function _setReserveFactor(uint256 newReserveFactor) external;
+	function _setKinkUtilizationRate(uint256 newKinkUtilizationRate) external;
+	function _setAdjustSpeed(uint256 newAdjustSpeed) external;
 	function _setBorrowTracker(address newBorrowTracker) external;
 }
