+++ src/interfaces/IImpermaxCallee.sol
@@ -1,6 +1,6 @@
 pragma solidity >=0.5.0;
 
 interface IImpermaxCallee {
-    function impermaxBorrow(address sender, address borrower, uint borrowAmount, bytes calldata data) external;
-    function impermaxRedeem(address sender, uint redeemAmount, bytes calldata data) external;
+    function impermaxBorrow(address sender, address borrower, uint256 borrowAmount, bytes calldata data) external;
+    function impermaxRedeem(address sender, uint256 redeemAmount, bytes calldata data) external;
 }
