+++ src/interfaces/IWETH.sol
@@ -3,7 +3,7 @@
 interface IWETH {
     function deposit() external payable;
 
-    function transfer(address to, uint value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
 
     function withdraw(uint) external;
 }
